And with that, I'm done with extractors for now. Just need to do some cleanup and documentation
