+++ src/main/java/autocadDrawingChecker/autocadData/elements/AutoCADPolyline.java
@@ -1,4 +1,4 @@
-package autocadDrawingChecker.autocadData;
+package autocadDrawingChecker.autocadData.elements;
 
 import autocadDrawingChecker.autocadData.elements.AutoCADElement;
